+++ frontend/javascripts/components/check_certificate_modal.tsx
